Refactoring
Internal structure refactoring, following the latest addition of
configuration tabs.

This has been done in a separate commit because the previous changes
were already very big by themselves.
